pyproject.toml
[project]
name = "sturmico"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "python-escpos[all]>=3.1",
]
